Create Inheritance
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed code for Person and a declaration for Student are provided for you in the editor. Observe that Student inherits all the properties of Person.

Complete the Student class by writing the following:

A Student class constructor, which has 4 parameters:
A string, firstName.
A string, lastName.
An integer, id .
An integer array (or vector) of test scores,scores .
A char calculate() method that calculates a Student object's average and returns the grade character representative of their calculated average:
O = 100-90
E = 90-80
A = 80-70
P = 70-55
D = 55-40
T = 40-0